feat: Add functionality to save an image file
This commit adds a new function `save_add_image` to the `lib.rs` file. This function allows the user to save the `ADD_IMAGE` to a file specified by the `file_path` parameter. The function uses the `File` struct from the `std::fs` module to create the file and write the image data to it.

The purpose of this change is to provide a convenient way for users to save the `ADD_IMAGE` to a file for further use or display.